# specs/001-budget-tracker/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Budget Tracker API
  description: REST API for managing checking account balance, virtual savings, recurring expenses, and pending payments
  version: 1.0.0
  contact:
    name: Budget Tracker Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
  /account:
    get:
      tags: [Account]
      summary: Get checking account balance
      description: Fetches the current checking account balance. Returns zero if no account exists.
      operationId: getAccount
      responses:
        '200':
          description: Account balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                balance: "3500.00"
                created_at: "2025-11-21T10:00:00Z"
                updated_at: "2025-11-21T15:30:00Z"
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags: [Account]
      summary: Update checking account balance
      description: Updates the checking account balance. Creates account if it doesn't exist.
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - balance
              properties:
                balance:
                  type: number
                  format: decimal
                  description: New account balance (can be negative for overdraft)
                  example: 3500.00
      responses:
        '200':
          description: Account balance updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List all recurring expenses
      description: Retrieves all monthly recurring expenses ordered by creation date
      operationId: listExpenses
      responses:
        '200':
          description: Expenses list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringExpense'
              example:
                - id: "770e8400-e29b-41d4-a716-446655440002"
                  description: "Rent"
                  amount: "1200.00"
                  created_at: "2025-11-21T10:10:00Z"
                  updated_at: "2025-11-21T10:10:00Z"
                - id: "770e8400-e29b-41d4-a716-446655440003"
                  description: "Netflix Subscription"
                  amount: "15.99"
                  created_at: "2025-11-21T10:11:00Z"
                  updated_at: "2025-11-21T10:11:00Z"
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Expenses]
      summary: Create new recurring expense
      description: Creates a new monthly recurring expense
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseInput'
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /expenses/{id}:
    put:
      tags: [Expenses]
      summary: Update recurring expense
      description: Updates description and/or amount of an existing expense
      operationId: updateExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseInput'
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpense'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags: [Expenses]
      summary: Delete recurring expense
      description: Permanently deletes an expense and releases its amount back to remaining budget
      operationId: deleteExpense
      parameters:
        - $ref: '#/components/parameters/ExpenseId'
      responses:
        '204':
          description: Expense deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /payments:
    get:
      tags: [Payments]
      summary: List all pending payments
      description: Retrieves all pending payments (both paid and unpaid) ordered by creation date
      operationId: listPayments
      responses:
        '200':
          description: Payments list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingPayment'
              example:
                - id: "880e8400-e29b-41d4-a716-446655440003"
                  description: "Car repair"
                  amount: "450.00"
                  is_paid: false
                  created_at: "2025-11-21T10:15:00Z"
                  updated_at: "2025-11-21T10:15:00Z"
                - id: "880e8400-e29b-41d4-a716-446655440004"
                  description: "Electric bill"
                  amount: "120.50"
                  is_paid: true
                  created_at: "2025-11-21T10:16:00Z"
                  updated_at: "2025-11-21T11:00:00Z"
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Payments]
      summary: Create new pending payment
      description: Creates a new pending payment (initially unpaid)
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PendingPaymentInput'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingPayment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /payments/{id}:
    put:
      tags: [Payments]
      summary: Update pending payment
      description: Updates description, amount, and/or paid status of a payment
      operationId: updatePayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - amount
                - is_paid
              properties:
                description:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: "Car repair"
                amount:
                  type: number
                  format: decimal
                  minimum: 0
                  example: 450.00
                is_paid:
                  type: boolean
                  description: Whether payment has been completed
                  example: false
      responses:
        '200':
          description: Payment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingPayment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags: [Payments]
      summary: Delete pending payment
      description: Permanently deletes a payment. If unpaid, releases its amount back to remaining budget.
      operationId: deletePayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '204':
          description: Payment deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  parameters:
...
    ExpenseId:
      name: id
      in: path
      required: true
      description: Recurring expense UUID
      schema:
        type: string
        format: uuid
    PaymentId:
      name: id
      in: path
      required: true
      description: Pending payment UUID
      schema:
        type: string
        format: uuid
  schemas:
    Account:
      type: object
      required:
        - id
        - balance
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique account identifier
        balance:
          type: string
          format: decimal
          description: Current account balance (can be negative)
          example: "3500.00"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    RecurringExpense:
      type: object
      required:
        - id
        - description
        - amount
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 200
          example: "Rent"
        amount:
          type: string
          format: decimal
          description: Monthly expense amount (must be >= 0)
          example: "1200.00"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecurringExpenseInput:
      type: object
      required:
        - description
        - amount
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 200
          description: Expense description
          example: "Rent"
        amount:
          type: number
          format: decimal
          minimum: 0
          description: Monthly expense amount (must be >= 0)
          example: 1200.00
    PendingPayment:
      type: object
      required:
        - id
        - description
        - amount
        - is_paid
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 200
          example: "Car repair"
        amount:
          type: string
          format: decimal
          description: Payment amount (must be >= 0)
          example: "450.00"
        is_paid:
          type: boolean
          description: Whether payment has been marked as paid
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PendingPaymentInput:
      type: object
      required:
        - description
        - amount
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 200
          description: Payment description
          example: "Car repair"
        amount:
          type: number
          format: decimal
          minimum: 0
          description: Payment amount (must be >= 0)
          example: 450.00
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Amount must be positive"
            details:
              type: object
              description: Additional error context
              additionalProperties: true
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Amount must be a positive number"
              details:
                field: "amount"
                value: -100
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              details:
                resource_id: "550e8400-e29b-41d4-a716-446655440000"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"